test(desktop): split icon tooltip smoke coverage
